Phase 1 files
Phase 1 is not related to codehub. It is the first part of the lab where we developed SRS and Project plan for "PESU App". After a few weeks, we had to exchange our documents with another team and continue working on the new project. We had to exchange it with a team who had "Code management and version control system" and we then developed codehub.